+++ NEWS.md
@@ -1,5 +1,12 @@
 # testthat 0.8.1.0.99
 
+* testhat now comes with a first version of a testing language for Behaviour
+  Driven Development.
+  The language is similiar to RSpec for Ruby or Mocha for JavaScript.
+  BDD tests read like sentences and it should thus be easier
+  to understand what the specification of a function/component is.
+  Refer to the `describe()` for further information and examples.
+
 # testthat 0.8.1
 
 * Better default environment for `test_check()` and `test_package()` which